Adding implementation for new CLI commands

+++ firmware/src/mqtt/mqtt_core/mqtt_core.h
@@ -224,7 +224,7 @@ typedef struct _mqttPublishPacket
     uint8_t packetIdentifierMSB;
 
     // Payload
-    uint8_t  payloadLength;
+    uint16_t payloadLength;
     uint8_t* payload;
 
     uint16_t totalLength;
